clean up link generation, moving TagletWriterImpl.linkSeeReferenceOutput to HtmlLinkTaglet
